blog_project/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.0.8
django-crispy-forms==1.9.1
django-six==1.0.4
et-xmlfile==1.0.1
Faker==4.1.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.6.0
isort==4.3.21
jdcal==1.4.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.1
oauthlib==3.1.0
openpyxl==3.0.3
packaging==20.3
pbr==5.4.5
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==1.32.0
text-unidecode==1.3
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.18
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
wrapt==1.11.2
xlwt==1.3.0
zipp==3.1.0
